+++ BambooTracker/misc.hpp
@@ -67,10 +67,10 @@ enum class SSGWaveFormType : int
 DECL_MAYBE_UNUSED
 static std::pair<int, Note> noteNumberToOctaveAndNote(int num)
 {
-	if (num < 0) return std::pair<int, Note>(0, Note::C);
+	if (num < 0) return std::make_pair(0, Note::C);
 
 	int oct = num / 12;
-	if (oct > 8) return std::pair<int, Note>(8, Note::B);
+	if (oct > 8) return std::make_pair(8, Note::B);
 
 	Note note;
 	switch (num % 12) {
@@ -88,7 +88,7 @@ static std::pair<int, Note> noteNumberToOctaveAndNote(int num)
 	case 11:	note = Note::B;		break;
 	}
 
-	return std::pair<int, Note>(oct, note);
+	return std::make_pair(oct, note);
 }
 
 DECL_MAYBE_UNUSED