+++ BambooTracker/opna_controller.hpp
@@ -212,6 +212,10 @@ public:
 	void setDetuneSSG(int ch, int pitch);
 	void setNoteSlideSSG(int ch, int speed, int seminote);
 	void setTransposeEffectSSG(int ch, int seminote);
+	void setToneNoiseMixSSG(int ch, int value);
+	void setNoisePitchSSG(int ch, int pitch);
+	void setHardEnvelopePeriod(int ch, bool high, int period);
+	void setAutoEnvelopeSSG(int ch, int shift, int shape);
 
 	// For state retrieve
 	void haltSequencesSSG(int ch);
@@ -265,6 +269,9 @@ private:
 	int sumNoteSldSSG_[3];
 	bool noteSldSSGSetFlag_;
 	int transposeSSG_[3];
+	int toneNoiseMixSSG_[3];
+	int noisePitchSSG_;
+	int hardEnvPeriodHighSSG_, hardEnvPeriodLowSSG_;
 	std::unique_ptr<int16_t[]> outputHistory_;
 	size_t outputHistoryIndex_;
 	std::unique_ptr<int16_t[]> outputHistoryReady_;
@@ -298,8 +305,9 @@ private:
 	/*----- Drum -----*/
 public:
 	// Key on-off
-	void keyOnDrum(int ch);
-	void keyOffDrum(int ch);
+	void setKeyOnFlagDrum(int ch);
+	void setKeyOffFlagDrum(int ch);
+	void updateKeyOnOffStatusDrum();
 
 	// Set volume
 	void setVolumeDrum(int ch, int volume);
@@ -314,6 +322,7 @@ public:
 	bool isMuteDrum(int ch);
 
 private:
+	uint8_t keyOnFlagDrum_, keyOffFlagDrum_;
 	int volDrum_[6], mVolDrum_, tmpVolDrum_[6];
 	/// bit0: right on/off
 	/// bit1: left on/off