+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - [#152] - Use RtAudio for sound processing ([#96]; thanks [@jpcima], [@OPNA2608] and others)
+- [#156] - Change key-on/off of multiple drum instruments to write to register at once (thanks [@ImATrackMan], [@ValleyBell])
 
 ### Fixed
 - Fix to draw the playing cursor when follow mode is turned off
@@ -14,10 +15,15 @@
 - [#154] - Fix position initialization when opening module (thanks [@papiezak])
 - [#96], [#152] - Reduce drawing cost of pattern editor and order list
 - Disable slider operation during playback
+- Fix corruption after `0Bxx`, `0Cxx` and `0Dxx` are set the position before current playback row
+- Fix incorrect square-mask pitch in square-masked saw and inversed saw
 
 [#152]: https://github.com/rerrahkr/BambooTracker/pull/152
 [#153]: https://github.com/rerrahkr/BambooTracker/issues/153
 [#154]: https://github.com/rerrahkr/BambooTracker/issues/154
+[#156]: https://github.com/rerrahkr/BambooTracker/issues/156
+
+[@ValleyBell]: https://github.com/ValleyBell
 
 ## v0.2.4 (2019-09-17)
 ### Changed